Kernel/aarch64: Move RP1 xHCI implementation to a separate file
This seems better than having all of this in a single file.
