+++ Kernel/CMakeLists.txt
@@ -483,7 +483,8 @@ elseif("${SERENITY_ARCH}" STREQUAL "aarch64")
         Arch/aarch64/RPi/InterruptController.cpp
         Arch/aarch64/RPi/Mailbox.cpp
         Arch/aarch64/RPi/MiniUART.cpp
-        Arch/aarch64/RPi/RP1.cpp
+        Arch/aarch64/RPi/RP1/RP1.cpp
+        Arch/aarch64/RPi/RP1/xHCI.cpp
         Arch/aarch64/RPi/SDHostController.cpp
         Arch/aarch64/RPi/Timer.cpp
         Arch/aarch64/RPi/Watchdog.cpp
